pyproject.toml
[tool.poetry]
name = "knowledge-graph-synth"
version = "0.1.0"
description = "A knowledge graph synthesis system that processes text documents, extracts entities and relationships, constructs knowledge graphs, and generates insights."
authors = ["Author <email@example.com>"]
readme = "README.md"
packages = [{include = "knowledge_graph_synth", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
spacy = "^3.8.4"
langdetect = "^1.0.9"
networkx = "^3.4.2"
pyvis = "^0.3.2"
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"
rich = "^13.9.4"
jinja2 = "^3.1.6"
diskcache = "^5.6.3"
aiohttp = "^3.11.13"
jsonschema = "^4.21.1"
google-genai = "^1.4.0"
plotly = "^6.0.0"
markdown = "^3.6"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.0.0"
black = "^25.1.0"
flake8 = "^7.1.2"
mypy = "^1.15.0"
isort = "^6.0.1"

[tool.poetry.scripts]
kgs = "knowledge_graph_synth.cli:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
